+++ src/data/duke.txt
@@ -1,3 +1,2 @@
-E | 1 | die  | 2023-01-01T09:30 | 2023-02-02T15:30
-T | 0 | adsnfjasd /by 2022-12-25 23:59
-D | 0 | christmas | 2022-12-25T23:59
+E | 0 | die  | 2023-01-01T09:30 | 2023-02-02T15:30
+T | 0 | hello
