+++ src/main/java/Exceptions/DukeException.java
@@ -1,6 +1,8 @@
+package Exceptions;
+
 public class DukeException extends Exception {
 
-    DukeException(String message) {
+    public DukeException(String message) {
         super(message);
     }
